requirements.txt
PyQt5
plyer
matplotlib
python-dotenv